Move sc_gl_graphics to mapping_core
Move sc_gl_graphics.h and sc_gl_graphics.cpp to mapping core, this is probably the last file that needs to be moved to enable standalone rendering
