+++ src/mapping_core/render/sc_gl_graphics.cpp
